requirements.txt
# Proton - Tree-based Agent Orchestration Platform
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Microsoft Agent Framework
agent-framework>=1.0.0b251001

# LLM Providers
openai>=1.10.0
anthropic>=0.18.0

# Azure support
azure-identity>=1.15.0

# MCP support
mcp>=0.1.0

# Vector databases (optional)
chromadb>=0.4.0
# pinecone-client>=3.0.0
# qdrant-client>=1.7.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
black>=24.1.0
ruff>=0.1.0
